+++ package/src/renderCaching.ts
@@ -1,6 +1,5 @@
 import type * as Runtime from "astro/compiler-runtime";
 import hashSum from "hash-sum";
-import { Cache } from "./cache.js";
 import { rootDebug } from "./debug.js";
 import type { AstroComponentFactory } from "astro/runtime/server/index.js";
 import type { SSRMetadata, SSRResult } from "astro";
@@ -8,9 +7,7 @@ import { runtime } from "./utils.js";
 import type { RenderDestination } from "astro/runtime/server/render/common.js";
 import type { PersistedMetadata } from "./renderFileStore.js";
 import { isDeepStrictEqual, types } from "node:util";
-import { computeEntryHash, getCurrentContext, makeContextTracking } from "./contextTracking.js";
-
-type CacheRenderingFn = (originalFn: typeof Runtime.createComponent) => typeof Runtime.createComponent;
+import { computeEntryHash, getCachingOptions, getCurrentContext, makeContextTracking } from "./contextTracking.js";
 
 const debug = rootDebug.extend('render-caching');
 
@@ -20,16 +17,9 @@ interface ExtendedSSRResult extends SSRResult {
   [ASSET_SERVICE_CALLS]: PersistedMetadata['assetServiceCalls'];
 }
 
-export const makeCaching = ({ cache, routeEntrypoints, componentHashes, ...cacheOptions }: {
-  cache: Cache,
-  root: string,
-  routeEntrypoints: string[],
-  componentHashes: Map<string, string>,
-  cacheComponents: false | 'in-memory' | 'persistent',
-  cachePages: boolean,
-}): CacheRenderingFn => (originalFn) => {
-  debug('Render caching called with:', { routeEntrypoints });
-
+(globalThis as any)[Symbol.for('@domain-expansion:astro-component-caching')] = (
+  originalFn: typeof Runtime.createComponent,
+): typeof Runtime.createComponent => {
   return function cachedCreateComponent(factoryOrOptions, moduleId, propagation) {
     const options = typeof factoryOrOptions === 'function'
       ? { factory: factoryOrOptions, moduleId, propagation } as Exclude<typeof factoryOrOptions, Function>
@@ -38,6 +28,8 @@ export const makeCaching = ({ cache, routeEntrypoints, componentHashes, ...cache
     const context = getCurrentContext();
 
     let cacheScope = options.moduleId || '';
+
+    const { componentHashes } = getCachingOptions();
 
     if (!options.moduleId || !componentHashes.has(options.moduleId)) {
       if (!context) return originalFn(options);
@@ -52,7 +44,6 @@ export const makeCaching = ({ cache, routeEntrypoints, componentHashes, ...cache
       cacheScope = `ccEntry:${ccRenderCall.id}:${ccRenderCall.hash}`;
     } else {
       const hash = componentHashes.get(options.moduleId)!;
-      debug('Creating cached component', { moduleId: options.moduleId, hash })
       cacheScope = hash;
     }
 
@@ -64,14 +55,27 @@ export const makeCaching = ({ cache, routeEntrypoints, componentHashes, ...cache
   }
 
   function cacheFn(cacheScope: string, factory: AstroComponentFactory, moduleId?: string): AstroComponentFactory {
+    const {
+      cache,
+      routeEntrypoints,
+      componentHashes,
+      componentsHaveSharedState,
+      ...cacheOptions
+    } = getCachingOptions();
+
     const isEntrypoint = routeEntrypoints.includes(moduleId!);
     const cacheParams: Record<'persist' | 'skipInMemory', boolean> = {
       persist: (
         (isEntrypoint && cacheOptions.cachePages)
         || (!isEntrypoint && cacheOptions.cacheComponents === 'persistent')
       ),
-      skipInMemory: isEntrypoint || cacheOptions.cacheComponents !== false,
+      skipInMemory: isEntrypoint || cacheOptions.cacheComponents === false,
     };
+
+    debug('Creating cached component', {
+      cacheScope, moduleId,
+      isEntrypoint, cacheParams,
+    });
 
     return async (result: ExtendedSSRResult, props, slots) => {
       const context = getCurrentContext();
@@ -90,10 +94,12 @@ export const makeCaching = ({ cache, routeEntrypoints, componentHashes, ...cache
       }
 
       // TODO: Handle edge-cases involving Object.defineProperty
-      const resolvedProps = Object.fromEntries((await Promise.all(
-        Object.entries(props)
-          .map(async ([key, value]) => [key, types.isProxy(value) ? undefined : await value])
-      )).filter((_key, value) => !!value));
+      const resolvedProps = Object.fromEntries(
+        (await Promise.all(
+          Object.entries(props)
+            .map(async ([key, value]) => [key, types.isProxy(value) ? undefined : await value])
+        ))
+          .filter(([_key, value]) => !!value));
 
       // We need to delete this because otherwise scopes from outside of a component can be globally
       // restricted to the inside of a child component through a slot and to support that the component
@@ -109,7 +115,11 @@ export const makeCaching = ({ cache, routeEntrypoints, componentHashes, ...cache
 
       const url = new URL(result.request.url);
 
-      const hash = hashSum([result.compressHTML, result.params, url.pathname, url.search, resolvedProps]);
+      const hash = hashSum(
+        isEntrypoint || componentsHaveSharedState
+          ? [moduleId, result.compressHTML, result.params, url.pathname, resolvedProps]
+          : [moduleId, result.compressHTML, resolvedProps]
+      );
       const cacheKey = `${cacheScope}:${hash}`;
 
       const { runIn: enterTrackingScope, collect: collectTracking } = makeContextTracking();
@@ -212,6 +222,11 @@ export const makeCaching = ({ cache, routeEntrypoints, componentHashes, ...cache
         if (currentHash !== hash) return null;
       }
 
+      for (const entry of cachedMetadata.renderEntryCalls) {
+        const currentHash = await computeEntryHash(entry.filePath);
+        if (currentHash !== entry.hash) return null;
+      }
+
       for (const { options, resultingAttributes } of cachedMetadata.assetServiceCalls) {
         debug('Replaying getImage call', { options });
         const result = await runtime.getImage(options);
@@ -220,11 +235,6 @@ export const makeCaching = ({ cache, routeEntrypoints, componentHashes, ...cache
           debug('Image call mismatch, bailing out of cache');
           return null;
         }
-      }
-
-      for (const entry of cachedMetadata.renderEntryCalls) {
-        const currentHash = await computeEntryHash(entry.filePath);
-        if (currentHash !== entry.hash) return null;
       }
 
       return cachedMetadata;